Some handy event tools.
