+++ soxs/tests/test_power_law.py
@@ -9,7 +9,8 @@ from soxs.instrument_registry import \
     get_instrument_from_registry
 from soxs.instrument import instrument_simulator, \
     RedistributionMatrixFile, AuxiliaryResponseFile
-from soxs.utils import write_spectrum, convert_rmf
+from soxs.utils import convert_rmf
+from soxs.events import write_spectrum
 from sherpa.astro.ui import load_user_model, add_user_pars, \
     load_pha, ignore, fit, set_model, set_stat, set_method, \
     covar, get_covar_results, set_covar_opt
